MEVA/MEVA/requirements.txt
pyyaml
future
tqdm
scipy
chumpy
numba
yacs
joblib
trimesh
pyrender
pytube
filterpy
h5py
scikit-image
opencv-python
torchvision
tensorboard
git+https://github.com/ZhengyiLuo/smplx.git
git+https://github.com/mkocabas/yolov3-pytorch.git

git+https://github.com/mkocabas/multi-person-tracker.git